ASTON-21 [refactor]: global refactoring
