+++ src/components/avatar/avatar.module.scss
@@ -6,7 +6,7 @@
   align-items: center;
   justify-content: center;
 
-  background: rgb(87, 84, 84);
+  background: rgb(87 84 84);
   border-radius: 50%;
   box-shadow: 0 2px 18px 0 rgb(88 88 88 / 35%);
