--- src/components/characterContainer/index.ts
@@ -1,1 +1,0 @@
-export * from './characterContainer'
