--- src/components/characterList/index.ts
@@ -1,1 +1,0 @@
-export * from './characterList'
