--- src/components/characterPage/index.ts
@@ -1,1 +1,0 @@
-export * from './characterPage'
