+++ src/components/charactersContainer/charactersContainer.module.scss
