+++ src/components/charactersList/charactersList.module.scss
