+++ src/components/charactersList/charactersList.tsx
@@ -3,7 +3,7 @@ import { FC } from 'react'
 import { Character } from '@/service/ResoursesService/CharactersApi'
 import clsx from 'clsx'
 
-import s from './characterList.module.scss'
+import s from './charactersList.module.scss'
 
 import { CharacterCard } from '../characterCard'
 
@@ -12,7 +12,7 @@ type CharacterListProps = {
   className?: string
 }
 
-export const CharacterList: FC<CharacterListProps> = ({ chars, className }) => {
+export const CharactersList: FC<CharacterListProps> = ({ chars, className }) => {
   return (
     <div className={clsx(s.character, className)}>
       <ul className={s.character__list}>
