+++ src/components/charactersList/index.ts
@@ -1,0 +1,1 @@
+export * from './charactersList'
