+++ src/page/character-page/character-page.module.scss
