+++ src/page/character-page/index.ts
@@ -1,0 +1,1 @@
+export * from './character-page'
