+++ src/page/character-page/lib/characterImageContainer.tsx
@@ -1,4 +1,4 @@
-import s from '../characterPage.module.scss'
+import s from '../character-page.module.scss'
 
 type Props = {
   img: string
