+++ src/page/search-page/index.ts
@@ -1,1 +1,1 @@
-export * from './search-page'
+export * from './search-page-container'
