+++ src/router/publicRouters.tsx
@@ -1,7 +1,8 @@
 import { RouteObject } from 'react-router-dom'
 
-import { urlPaths } from '@/enum'
-import { SearchPage } from '@/page/search-page'
+import { urlPaths } from '@/enums'
+import { CharacterPageContainer } from '@/page/character-page/character-page-container'
+import { SearchPageContainer } from '@/page/search-page'
 import { SignInPageContainer } from '@/page/sign-in-page'
 import { SignUpPageContainer } from '@/page/sign-up-page'
 
@@ -19,16 +20,16 @@ export const publicRouters: RouteObject[] = [
     path: urlPaths.error,
   },
   {
-    element: <h1>Main</h1>,
+    element: <SearchPageContainer />,
     path: urlPaths.root,
   },
 
   {
-    element: <h1>Character</h1>,
+    element: <CharacterPageContainer />,
     path: urlPaths.chapterId,
   },
   {
-    element: <SearchPage />,
+    element: <SearchPageContainer />,
     path: urlPaths.search,
   },
 ]
